202307/requirements.txt
PyYAML==6.0
accelerate @ git+https://github.com/huggingface/accelerate.git@7b4d12623a1847bf7f749d6824017683922339eb
bitsandbytes @ file:///home/acf15478nt/nvidia_hackathon/202307/text_generation/bitsandbytes
datasets==2.12.0
deepspeed==0.9.2
einops==0.6.1
evaluate==0.4.0
peft==0.3.0
protobuf==4.23.3
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.99
torch==2.0.1
transformers @ git+https://github.com/huggingface/transformers@de9255de27abfcae4a1f816b904915f0b1e23cd9
fire==0.5.0
